server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pillow>=10.3.0
aiosqlite>=0.19.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.1.0
websockets>=12.0
# Pulsar dependencies (should be installed in sage environment)
# torch
# diffusers
# tqdm
# bitarray
# pypng
# reedsolo
# reedmuller
